Fixed security bugs for view orders/products,delete orders,users,prod…

+++ backend/deleteUser.php
@@ -1,9 +1,18 @@
 <?php
 include_once 'mappers/userMapper.php';
-if (isset($_GET['id'])) {
-    $userId = $_GET['id'];
-    $mapper = new UserMapper();
-    $mapper->deleteUser($userId);
-    header("Location:../pages/Dashboard.php");
+session_start();
+if($_SESSION['role']==0){
+    header("Location:../index.php");
 }
+else{
+
+    if (isset($_GET['id'])) {
+        $userId = $_GET['id'];
+        $mapper = new UserMapper();
+        $mapper->deleteUser($userId);
+        header("Location:../pages/Dashboard.php");
+    }
+    
+}
+
